+++ src/main/java/org/broadinstitute/hellbender/tools/spark/utils/LongHopscotchSet.java
@@ -264,7 +264,7 @@ public final class LongHopscotchSet implements Serializable {
             buckets[bucketIndex] = 0;
             status[bucketIndex] = 0;
             if (predecessorIndex != NO_ELEMENT_INDEX) { // fix up offset of previous element in chain if there is one
-                status[predecessorIndex] -= getOffset(predecessorIndex);
+                status[predecessorIndex] -= (byte) getOffset(predecessorIndex);
             }
         } else {
             // move the item at the end of the chain into the hole we're creating by deleting this entry
@@ -277,7 +277,7 @@ public final class LongHopscotchSet implements Serializable {
             }
             buckets[bucketIndex] = buckets[nextIndex];
             buckets[nextIndex] = 0;
-            status[prevIndex] -= getOffset(prevIndex);
+            status[prevIndex] -= (byte) getOffset(prevIndex);
         }
         size -= 1;
     }
@@ -299,7 +299,7 @@ public final class LongHopscotchSet implements Serializable {
 
         // if the new entry lies downstream of the current chain end, just link it in
         if (offsetToEmpty > offsetToEndOfChain) {
-            status[endOfChainIndex] += offsetToEmpty - offsetToEndOfChain;
+            status[endOfChainIndex] += (byte) (offsetToEmpty - offsetToEndOfChain);
         } else {
             linkIntoChain(bucketIndex, emptyBucketIndex);
         }
@@ -317,7 +317,7 @@ public final class LongHopscotchSet implements Serializable {
             offsetToEmpty -= offset;
         }
         offset -= offsetToEmpty;
-        status[tmpIndex] -= offset;
+        status[tmpIndex] -= (byte) offset;
         status[emptyBucketIndex] = (byte) offset;
     }
 
@@ -343,7 +343,7 @@ public final class LongHopscotchSet implements Serializable {
             buckets[emptyBucketIndex] = buckets[nextIndex];
             buckets[nextIndex] = 0;
             status[nextIndex] = 0;
-            status[prevIndex] -= getOffset(prevIndex);
+            status[prevIndex] -= (byte) getOffset(prevIndex);
             emptyBucketIndex = nextIndex;
         }
     }
@@ -410,9 +410,9 @@ public final class LongHopscotchSet implements Serializable {
         int toEmptyDistance = getIndexDiff(bucketToMoveIndex, emptyBucketIndex);
         int nextOffset = getOffset(bucketToMoveIndex);
         if (nextOffset == 0 || nextOffset > toEmptyDistance) {
-            status[predecessorBucketIndex] += toEmptyDistance;
+            status[predecessorBucketIndex] += (byte) toEmptyDistance;
         } else {
-            status[predecessorBucketIndex] += nextOffset;
+            status[predecessorBucketIndex] += (byte) nextOffset;
             toEmptyDistance -= nextOffset;
             predecessorBucketIndex = getIndex(bucketToMoveIndex, nextOffset);
             while ((nextOffset = getOffset(predecessorBucketIndex)) != 0 && nextOffset < toEmptyDistance) {
